Update EmployeeUserInterface.java
This is the final form of the EmployeeUserInterface class, with a fully implemented calculate button with some basic error handling. 
When the user clicks the 'Calculate' button, the program will scrape whatever information has been entered into the JTextFields and generate an Employee object, and then use that Employee object's raiseSalary() method to generate a raised salary and display that information in a pop-up window. If the information in the text-fields is insufficient to generate a proper Employee object, a warning window will pop up and scold the user.